refactor(solana): harden wallet integration and extract transaction builders
- Replace contractAddress truthiness checks with explicit NATIVE_SOL_ADDRESS
  comparison in transactionBuilder and client (handles null from API correctly)
- Add NATIVE_SOL_ADDRESS constant to constants.ts; replace all hardcoded
  system address strings in client.ts with it
- Extract refundTransactionBuilder and redeemSolverTransactionBuilder from
  client.ts into transactionBuilder.ts; client now only handles sign/send/confirm
- Remove useSolanaWalletStore: signMessage is now read from the live adapter
  directly in walletLoginContext (same pattern as sendTransaction)
- Fix WalletNotConnectedError in useHTLCWriteClient: call adapter.sendTransaction
  directly instead of solanaSendTransaction from WalletProvider context
- Fix gas estimation: correct lpAddress → srcLpAddress, hashlock type,
  contractAddress null-safety, and replace as any with proper dummy params
- Rename phtlcTransactionBuilder export to userLockTransactionBuilder
