--- apps/app/stores/solanaWalletStore.ts
@@ -1,11 +1,0 @@
-import { create } from 'zustand'
-
-interface SolanaWalletState {
-  signMessage: ((message: Uint8Array) => Promise<Uint8Array>) | undefined
-  setSignMessage: (fn: SolanaWalletState['signMessage']) => void
-}
-
-export const useSolanaWalletStore = create<SolanaWalletState>()((set) => ({
-  signMessage: undefined,
-  setSignMessage: (fn) => set({ signMessage: fn }),
-}))
