+++ packages/blockchains/solana/src/constants.ts
@@ -1,0 +1,1 @@
+export const NATIVE_SOL_ADDRESS = '11111111111111111111111111111111'
