+++ packages/blockchains/solana/src/index.ts
@@ -8,8 +8,8 @@ export { SolanaHTLCClient } from './client.js'
 export type { SolanaHTLCClientConfig, SolanaSigner } from './types.js'
 export { deriveKeyFromSolanaWallet } from './login/index.js'
 export type { SolanaWalletLike } from './login/index.js'
-export { userLockTransactionBuilder as phtlcTransactionBuilder } from './transactionBuilder.js'
-export type { UserLockParams as PhtlcParams } from './transactionBuilder.js'
+export { userLockTransactionBuilder, refundTransactionBuilder, redeemSolverTransactionBuilder } from './transactionBuilder.js'
+export type { UserLockParams, RefundTxParams, RedeemSolverTxParams } from './transactionBuilder.js'
 export { TrainHtlc } from './idl/trainHtlc.js'
 export const TRAIN_HTLC_PROGRAM_ID = '6zasug6x5AY93zNVjPZPGoqQfdTBd3C1w6CU9NDKtNH8'
