+++ packages/blockchains/solana/src/transactionBuilder.ts
@@ -1,9 +1,21 @@
-import { Connection, PublicKey, Transaction } from "@solana/web3.js"
+import { Connection, PublicKey, Transaction, TransactionInstruction } from "@solana/web3.js"
 import { BN, Idl, Program } from "@coral-xyz/anchor"
-import { UserLockParams as SdkUserLockParams } from '@train-protocol/sdk'
-import { NATIVE_SOL_ADDRESS } from './types.js'
+import { UserLockParams as SdkUserLockParams, RefundParams, RedeemSolverParams } from '@train-protocol/sdk'
+import { NATIVE_SOL_ADDRESS } from './constants.js'
 
 export type UserLockParams = SdkUserLockParams & {
+    connection: Connection
+    program: Program<Idl>
+    walletPublicKey: PublicKey
+}
+
+export type RefundTxParams = RefundParams & {
+    connection: Connection
+    program: Program<Idl>
+    walletPublicKey: PublicKey
+}
+
+export type RedeemSolverTxParams = RedeemSolverParams & {
     connection: Connection
     program: Program<Idl>
     walletPublicKey: PublicKey
@@ -22,6 +34,13 @@ function toBaseUnits(amount: string, decimals: number): BN {
     const [int, frac = ''] = amount.split('.')
     const fracPadded = frac.padEnd(decimals, '0').slice(0, decimals)
     return new BN(int + fracPadded)
+}
+
+function secretToBuffer(secret: string | bigint): Buffer {
+    if (typeof secret === 'bigint') {
+        return Buffer.from(secret.toString(16).padStart(64, '0'), 'hex')
+    }
+    return Buffer.from(secret.replace('0x', ''), 'hex')
 }
 
 export const userLockTransactionBuilder = async (params: UserLockParams): Promise<TransactionResult> => {
@@ -97,6 +116,132 @@ export const userLockTransactionBuilder = async (params: UserLockParams): Promis
             .transaction()
 
         tx.add(lockTx)
+    }
+
+    const blockHash = await connection.getLatestBlockhash()
+    tx.recentBlockhash = blockHash.blockhash
+    tx.lastValidBlockHeight = blockHash.lastValidBlockHeight
+    tx.feePayer = walletPublicKey
+
+    return { transaction: tx, blockhash: blockHash.blockhash, lastValidBlockHeight: blockHash.lastValidBlockHeight }
+}
+
+export const refundTransactionBuilder = async (params: RefundTxParams): Promise<TransactionResult> => {
+    const { connection, program, walletPublicKey } = params
+    const hashlockBuffer = Buffer.from(params.id.replace('0x', ''), 'hex')
+    const hashlockArray = Array.from(hashlockBuffer)
+
+    const [userLockPda] = PublicKey.findProgramAddressSync(
+        [Buffer.from("user_lock"), hashlockBuffer],
+        program.programId
+    )
+
+    let refundIx: TransactionInstruction
+    if (params.sourceAsset.contractAddress && params.sourceAsset.contractAddress !== NATIVE_SOL_ADDRESS) {
+        const { getAssociatedTokenAddress, TOKEN_PROGRAM_ID, ASSOCIATED_TOKEN_PROGRAM_ID } = await import('@solana/spl-token')
+        const tokenMint = new PublicKey(params.sourceAsset.contractAddress)
+        const senderTokenAccount = await getAssociatedTokenAddress(tokenMint, walletPublicKey)
+        const [vault] = PublicKey.findProgramAddressSync(
+            [Buffer.from("vault"), hashlockBuffer],
+            program.programId
+        )
+
+        refundIx = await program.methods
+            .refundUserToken(hashlockArray)
+            .accounts({
+                caller: walletPublicKey,
+                userLock: userLockPda,
+                sender: walletPublicKey,
+                tokenMint,
+                vault,
+                senderTokenAccount,
+                tokenProgram: TOKEN_PROGRAM_ID,
+                associatedTokenProgram: ASSOCIATED_TOKEN_PROGRAM_ID,
+            })
+            .instruction()
+    } else {
+        refundIx = await program.methods
+            .refundUserSol(hashlockArray)
+            .accounts({
+                caller: walletPublicKey,
+                userLock: userLockPda,
+                sender: walletPublicKey,
+            })
+            .instruction()
+    }
+
+    const closeIx = await program.methods
+        .closeUserLock(hashlockArray)
+        .accounts({ caller: walletPublicKey, userLock: userLockPda })
+        .instruction()
+
+    const { blockhash, lastValidBlockHeight } = await connection.getLatestBlockhash()
+    const tx = new Transaction()
+    tx.recentBlockhash = blockhash
+    tx.lastValidBlockHeight = lastValidBlockHeight
+    tx.feePayer = walletPublicKey
+    tx.add(refundIx, closeIx)
+
+    return { transaction: tx, blockhash, lastValidBlockHeight }
+}
+
+export const redeemSolverTransactionBuilder = async (params: RedeemSolverTxParams): Promise<TransactionResult> => {
+    const { connection, program, walletPublicKey } = params
+    const hashlockBuffer = Buffer.from(params.id.replace('0x', ''), 'hex')
+    const hashlockArray = Array.from(hashlockBuffer)
+    const secretArray = Array.from(secretToBuffer(params.secret))
+    const lockIndex = params.index ?? 1
+
+    const indexBuffer = Buffer.alloc(8)
+    indexBuffer.writeBigUInt64LE(BigInt(lockIndex))
+
+    const [solverLockPda] = PublicKey.findProgramAddressSync(
+        [Buffer.from("solver_lock"), hashlockBuffer, indexBuffer],
+        program.programId
+    )
+
+    const solverLockAccount = await (program.account as any).solverLock.fetch(solverLockPda)
+    const rewardRecipient = new PublicKey(solverLockAccount.rewardRecipient)
+    const recipient = params.destinationAddress ? new PublicKey(params.destinationAddress) : walletPublicKey
+
+    let tx: Transaction
+    if (params.sourceAsset.contractAddress && params.sourceAsset.contractAddress !== NATIVE_SOL_ADDRESS) {
+        const { getAssociatedTokenAddress, TOKEN_PROGRAM_ID, ASSOCIATED_TOKEN_PROGRAM_ID } = await import('@solana/spl-token')
+        const tokenMint = new PublicKey(params.sourceAsset.contractAddress)
+        const [vault] = PublicKey.findProgramAddressSync(
+            [Buffer.from("vault"), hashlockBuffer, indexBuffer],
+            program.programId
+        )
+        const recipientTokenAccount = await getAssociatedTokenAddress(tokenMint, recipient)
+        const rewardRecipientTokenAccount = await getAssociatedTokenAddress(tokenMint, rewardRecipient)
+        const callerTokenAccount = await getAssociatedTokenAddress(tokenMint, walletPublicKey)
+
+        tx = await program.methods
+            .redeemSolverToken(hashlockArray, lockIndex, secretArray)
+            .accounts({
+                caller: walletPublicKey,
+                solverLock: solverLockPda,
+                recipient,
+                rewardRecipient,
+                tokenMint,
+                vault,
+                recipientTokenAccount,
+                rewardRecipientTokenAccount,
+                callerTokenAccount,
+                tokenProgram: TOKEN_PROGRAM_ID,
+                associatedTokenProgram: ASSOCIATED_TOKEN_PROGRAM_ID,
+            })
+            .transaction()
+    } else {
+        tx = await program.methods
+            .redeemSolverSol(hashlockArray, lockIndex, secretArray)
+            .accounts({
+                caller: walletPublicKey,
+                solverLock: solverLockPda,
+                recipient,
+                rewardRecipient,
+            })
+            .transaction()
     }
 
     const blockHash = await connection.getLatestBlockhash()
